requirements.txt
requests>=2.28.0
python-dateutil>=2.8.0
dataclasses-json>=0.5.0
pytest>=7.0.0
pytest-cov>=4.0.0
python-dotenv>=1.0.0 